rename object_to_to_thread to object_store_adapter
