+++ src/lib.rs
@@ -1,4 +1,4 @@
 mod io_uring_local;
-mod object_store_to_thread;
+mod object_store_adapter;
 mod operation;
 mod operation_future;
