+++ src/object_store_adapter.rs
